+++ codegen/src/utils/meta_item_ext.rs
@@ -4,7 +4,7 @@ use syntax::symbol::Symbol;
 pub trait MetaItemExt {
     fn name_value(&self) -> Option<(&Symbol, &Lit)>;
     fn str_lit(&self) -> Option<&Symbol>;
-    fn int_lit(&self) -> Option<u64>;
+    fn int_lit(&self) -> Option<u128>;
 }
 
 impl MetaItemExt for NestedMetaItem {
@@ -22,7 +22,7 @@ impl MetaItemExt for NestedMetaItem {
         })
     }
 
-    fn int_lit(&self) -> Option<u64> {
+    fn int_lit(&self) -> Option<u128> {
         self.literal().and_then(|lit| match lit.node {
             LitKind::Int(n, _) => Some(n),
             _ => None,
